requirements.txt
fastapi~=0.101
uvicorn~=0.23
pydantic~=2.2
pydantic-settings==2.0.3
python-dotenv~=1.0
pymongo~=4.4
Jinja2~=3.1
python-multipart==0.0.6
Unidecode~=1.3
dateparser~=1.1
loguru~=0.7
Pillow==10.0.0
requests~=2.31
PyJWT~=2.8
pytest==7.4.0
pytest-asyncio==0.21.0
python-Levenshtein==0.26.1
wordcloud==1.9.4
